+++ SilverLite/drv_osd.h
@@ -12,7 +12,7 @@ void osd_print(int row, int col, const char *s);
 void osd_erase(int row, int col, int count);
 void osd_clear();
 void osd_refresh(void);
-uint8_t osd_get_max_rows();
+int  osd_get_max_rows();
 
 #if defined(__cplusplus)
 }
